+++ py_src/fusion/fusion.py
@@ -26,12 +26,11 @@ from fusion.fusion_types import Types
 from fusion.product import Product
 from fusion.report import Report
 
+from .embeddings_utils import _format_full_index_response, _format_summary_index_response
 from .exceptions import APIResponseError, FileFormatError
 from .fusion_filesystem import FusionHTTPFileSystem
 from .utils import (
     RECOGNIZED_FORMATS,
-    _format_full_index_response,
-    _format_summary_index_response,
     cpu_count,
     csv_to_table,
     distribution_to_filename,
@@ -57,7 +56,7 @@ from .utils import (
 if TYPE_CHECKING:
     import fsspec
     import requests
-    from opensearchpy import OpenSearch
+    from opensearchpy import AsyncOpenSearch, OpenSearch
 
     from .types import PyArrowFilterT
 
@@ -216,13 +215,19 @@ class Fusion:
 
         return catalog
 
-    def get_fusion_filesystem(self) -> FusionHTTPFileSystem:
-        """Creates Fusion Filesystem.
+    def get_fusion_filesystem(self, **kwargs: Any) -> FusionHTTPFileSystem:
+        """Retrieve Fusion file system instance.
+
+        Note: This function always returns a reference to the exact same FFS instance since
+        an FFS instance is based off the FusionCredentials object.
 
         Returns: Fusion Filesystem
 
         """
-        return FusionHTTPFileSystem(client_kwargs={"root_url": self.root_url, "credentials": self.credentials})
+        as_async = kwargs.get("asynchronous", False)
+        return FusionHTTPFileSystem(
+            asynchronous=as_async, client_kwargs={"root_url": self.root_url, "credentials": self.credentials}
+        )
 
     def list_catalogs(self, output: bool = False) -> pd.DataFrame:
         """Lists the catalogs available to the API account.
@@ -1204,6 +1209,7 @@ class Fusion:
         catalog: str | None = None,
         distribution: str = "parquet",
         show_progress: bool = True,
+        multipart: bool = True,
         return_paths: bool = False,
         chunk_size: int = 5 * 2**20,
         from_date: str | None = None,
@@ -1221,6 +1227,7 @@ class Fusion:
             catalog (str, optional): A catalog identifier. Defaults to 'common'.
             distribution (str, optional): A distribution type, e.g. a file format or raw
             show_progress (bool, optional): Display a progress bar during data download Defaults to True.
+            multipart (bool, optional): Is multipart upload.
             return_paths (bool, optional): Return paths and success statuses of the downloaded files.
             chunk_size (int, optional): Maximum chunk size.
             from_date (str, optional): start of the data date range contained in the distribution,
@@ -1250,7 +1257,7 @@ class Fusion:
             data_map_df,
             parallel=False,
             n_par=1,
-            multipart=False,
+            multipart=multipart,
             chunk_size=chunk_size,
             show_progress=show_progress,
             from_date=from_date,
@@ -2516,3 +2523,60 @@ class Fusion:
             root_url=self.root_url,
             credentials=self.credentials,
         )
+
+    def get_async_fusion_vector_store_client(self, knowledge_base: str, catalog: str | None = None) -> AsyncOpenSearch:
+        """Returns Fusion Embeddings Search client.
+
+        Args:
+            knowledge_base (str): Knowledge base (dataset) identifier.
+            catalog (str | None, optional): A catalog identifier. Defaults to 'common'.
+
+        Returns:
+            OpenSearch: Fusion Embeddings Search client.
+
+        """
+        from opensearchpy import AsyncOpenSearch
+
+        from fusion.embeddings import FusionAsyncHttpConnection
+
+        catalog = self._use_catalog(catalog)
+        return AsyncOpenSearch(
+            connection_class=FusionAsyncHttpConnection,
+            catalog=catalog,
+            knowledge_base=knowledge_base,
+            root_url=self.root_url,
+            credentials=self.credentials,
+        )
+    
+    def list_datasetmembers_distributions(
+        self,
+        dataset: str,
+        catalog: str | None = None,
+    ) -> pd.DataFrame:
+        """List the distributions of dataset members.
+
+        Args:
+            dataset (str): Dataset identifier.
+            catalog (str | None, optional): A catalog identifier. Defaults to 'common'.
+            output (bool, optional): If True then print the dataframe. Defaults to False.
+
+        Returns:
+            pd.DataFrame: A dataframe with a row for each dataset member distribution.
+
+        """
+        catalog = self._use_catalog(catalog)
+
+        url = f"{self.root_url}catalogs/{catalog}/datasets/changes?datasets={dataset}"
+
+        resp = self.session.get(url)
+        dists = resp.json()["datasets"][0]["distributions"]
+
+        data = []
+        for dist in dists:
+            values = dist.get("values")
+            member_id = values[5]
+            member_format = values[6]
+            data.append((member_id, member_format))
+
+        members_df = pd.DataFrame(data, columns=["identifier", "format"])
+        return members_df