+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.0.9] - 2025-03-11
+* Adding async functionality to FusionHTTPFileSystem
+
+## [2.0.8] - 2025-03-04
+* GZ supported distribution type
+* from bytes to handle multipart upload for large files
+
 ## [2.0.7] - 2025-02-03
 * Embeddings API functionality